# specs/openapi.yaml
openapi: 3.1.0
info:
  title: AIOps Console API
  version: 0.1.0-alpha
  description: |
    AI 原生后台管理系统 API，覆盖 Prompt 管理、Agent 编排、知识库、模型路由、
    对话分析与评估套件六大领域。

    遵循 Agentic Engineering 哲学：Spec 即契约，类型即文档。
  contact:
    name: AIOps Console Team
    url: https://github.com/aiops/console
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.aiops.local/api/v1
    description: Production
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      summary: 列出 Agent
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Agent 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentOut'
    post:
      tags: [Agents]
      summary: 创建 Agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /agents/{agent_id}/execute:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      tags: [Agents]
      summary: 执行 Agent
      operationId: executeAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  description: 用户输入
                stream:
                  type: boolean
                  default: false
                  description: 是否以流式返回
      responses:
        '200':
          description: 执行结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /workflows:
    get:
      tags: [Workflows]
      summary: 列出工作流
      operationId: listWorkflows
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 工作流列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowOut'
    post:
      tags: [Workflows]
      summary: 创建工作流
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDef'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /workflows/{workflow_id}/execute:
    parameters:
      - $ref: '#/components/parameters/WorkflowId'
    post:
      tags: [Workflows]
      summary: 执行工作流
      operationId: executeWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                stream:
                  type: boolean
                  default: false
      responses:
        '200':
          description: 执行追踪
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionTrace'
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /knowledge-bases:
    get:
      tags: [KnowledgeBases]
      summary: 列出知识库
      operationId: listKnowledgeBases
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 知识库列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/KnowledgeBaseOut'
    post:
      tags: [KnowledgeBases]
      summary: 创建知识库
      operationId: createKnowledgeBase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeBaseCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /knowledge-bases/{kb_id}/documents:
    parameters:
      - $ref: '#/components/parameters/KbId'
    post:
      tags: [KnowledgeBases]
      summary: 上传文档到知识库
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 文档文件（PDF/TXT/MD/DOCX，最大 50MB）
      responses:
        '202':
          description: 上传成功，文档进入处理队列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentOut'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /knowledge-bases/{kb_id}/search:
    parameters:
      - $ref: '#/components/parameters/KbId'
    post:
      tags: [KnowledgeBases]
      summary: 语义搜索知识库
      operationId: searchKnowledgeBase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                top_k:
                  type: integer
                  default: 5
                  minimum: 1
                  maximum: 20
                threshold:
                  type: number
                  format: float
                  default: 0.7
                  minimum: 0
                  maximum: 1
                  description: 相似度阈值，低于此分数的结果被过滤
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /models:
    get:
      tags: [Models]
      summary: 列出模型配置
      operationId: listModels
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 模型配置列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelConfigOut'
    post:
      tags: [Models]
      summary: 创建模型配置
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelConfigCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelConfigOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /models/{model_alias}/chat:
    parameters:
      - $ref: '#/components/parameters/ModelAlias'
    post:
      tags: [Models]
      summary: 聊天补全
      operationId: chatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: 聊天补全响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analytics/conversations:
    get:
      tags: [Analytics]
      summary: 列出对话记录
      operationId: listConversations
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: agent_id
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 对话记录列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /analytics/dashboard:
    get:
      tags: [Analytics]
      summary: 获取仪表盘指标
      operationId: getDashboard
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 仪表盘指标
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardMetrics'
  /evals:
    get:
      tags: [Evals]
      summary: 列出评估任务
      operationId: listEvals
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 评估任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/EvalRunOut'
    post:
      tags: [Evals]
      summary: 创建评估任务
      operationId: createEval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvalRunCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRunOut'
        '422':
          $ref: '#/components/responses/ValidationError'
  /evals/{eval_id}/run:
    parameters:
      - $ref: '#/components/parameters/EvalId'
    post:
      tags: [Evals]
      summary: 执行评估任务
      operationId: runEval
      responses:
        '202':
          description: 评估任务已开始执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRunOut'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
...
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: integer
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: integer
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: integer
    ModelAlias:
      name: model_alias
      in: path
      required: true
      schema:
        type: string
    EvalId:
      name: eval_id
      in: path
      required: true
      schema:
        type: integer
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
...
    AgentCreate:
      type: object
      required: [name, system_prompt, model_alias]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 512
        system_prompt:
          type: string
          maxLength: 65536
        model_alias:
          type: string
          description: 关联的模型配置别名
        tools:
          type: array
          items:
            type: string
          default: []
          description: Agent 可调用的工具名称列表
    AgentOut:
      type: object
      required: [id, name, system_prompt, model_alias, tools, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        system_prompt:
          type: string
        model_alias:
          type: string
        tools:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    WorkflowDef:
      type: object
      required: [name, nodes, entry_node]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/AgentNode'
          minItems: 1
          maxItems: 50
          description: 节点列表（DAG 最大 50 节点）
        entry_node:
          type: string
          description: 入口节点 ID
        max_rounds:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: 单次执行最大轮数
    WorkflowOut:
      type: object
      required: [id, name, nodes, entry_node, max_rounds, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/AgentNode'
        entry_node:
          type: string
        max_rounds:
          type: integer
        created_at:
          type: string
          format: date-time
    ExecutionResult:
      type: object
      required: [trace_id, status, output, token_usage, latency_ms]
      properties:
        trace_id:
          type: string
          description: 执行追踪 ID
        status:
          type: string
          enum: [success, failed, timeout]
        output:
          type: object
          description: 执行输出（自由结构）
        token_usage:
          type: object
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        latency_ms:
          type: integer
          description: 执行耗时（毫秒）
    ExecutionTrace:
      type: object
      required: [trace_id, workflow_id, status, spans, token_usage, latency_ms, created_at]
      properties:
        trace_id:
          type: string
        workflow_id:
          type: integer
        status:
          type: string
          enum: [running, success, failed, timeout]
        spans:
          type: array
          items:
            type: object
            properties:
              node_id:
                type: string
              agent_id:
                type: integer
              input:
                type: object
              output:
                type: object
              status:
                type: string
              start_time:
                type: string
                format: date-time
              end_time:
                type: string
                format: date-time
              token_usage:
                type: object
        token_usage:
          type: object
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        latency_ms:
          type: integer
        created_at:
          type: string
          format: date-time
    KnowledgeBaseCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          maxLength: 512
        embedding_model:
          type: string
          default: text-embedding-3-small
          description: 嵌入模型名称
        chunk_size:
          type: integer
          minimum: 256
          maximum: 2048
          default: 512
          description: 分块大小（字符数）
        chunk_overlap:
          type: integer
          minimum: 0
          maximum: 512
          default: 50
          description: 分块重叠（字符数）
    KnowledgeBaseOut:
      type: object
      required: [id, name, embedding_model, chunk_size, chunk_overlap, doc_count, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        embedding_model:
          type: string
        chunk_size:
          type: integer
        chunk_overlap:
          type: integer
        doc_count:
          type: integer
          description: 文档总数
        created_at:
          type: string
          format: date-time
    DocumentOut:
      type: object
      required: [id, kb_id, filename, file_size, chunk_count, status, created_at]
      properties:
        id:
          type: integer
        kb_id:
          type: integer
        filename:
          type: string
        file_size:
          type: integer
          description: 文件大小（字节，最大 50MB）
        chunk_count:
          type: integer
          description: 分块数量
        status:
          type: string
          enum: [pending, processing, done, error]
        created_at:
          type: string
          format: date-time
    SearchResult:
      type: object
      required: [chunk_id, doc_id, content, score, metadata]
      properties:
        chunk_id:
          type: string
        doc_id:
          type: integer
        content:
          type: string
          description: 文本块内容
        score:
          type: number
          format: float
          description: 相似度分数（0-1）
        metadata:
          type: object
          description: 文档元数据
          additionalProperties: true
    ModelConfigCreate:
      type: object
      required: [alias, provider_name, model_id]
      properties:
        alias:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          description: 模型别名（用于路由调用）
        provider_name:
          type: string
          enum: [openai, anthropic, azure, local]
        model_id:
          type: string
          description: 供应商模型标识
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 8192
          default: 4096
        cost_per_1k_input:
          type: number
          format: float
          minimum: 0
          description: 每 1k 输入 token 成本（美元）
        cost_per_1k_output:
          type: number
          format: float
          minimum: 0
          description: 每 1k 输出 token 成本（美元）
        routing_strategy:
          type: string
          enum: [direct, round_robin, least_cost, latency]
          default: direct
        fallback_models:
          type: array
          items:
            type: string
          default: []
          description: 备选模型别名列表
        quota_daily:
          type: integer
          default: 1000000
          description: 每日 token 配额
    ModelConfigOut:
      type: object
      required: [id, alias, provider_name, model_id, temperature, max_tokens, routing_strategy, fallback_models, quota_daily, enabled, created_at]
      properties:
        id:
          type: integer
        alias:
          type: string
        provider_name:
          type: string
          enum: [openai, anthropic, azure, local]
        model_id:
          type: string
        temperature:
          type: number
          format: float
        max_tokens:
          type: integer
        cost_per_1k_input:
          type: number
          format: float
        cost_per_1k_output:
          type: number
          format: float
        routing_strategy:
          type: string
          enum: [direct, round_robin, least_cost, latency]
        fallback_models:
          type: array
          items:
            type: string
        quota_daily:
          type: integer
        enabled:
          type: boolean
          description: 是否启用
        created_at:
          type: string
          format: date-time
...
    ChatRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          minItems: 1
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer
          minimum: 1
          maximum: 8192
        stream:
          type: boolean
          default: false
          description: 是否流式返回
    ChatResponse:
      type: object
      required: [id, content, model, usage, latency_ms]
      properties:
        id:
          type: string
          description: 响应 ID
        content:
          type: string
          description: 补全内容
        model:
          type: string
          description: 实际使用的模型标识
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        latency_ms:
          type: integer
    ConversationOut:
      type: object
      required: [id, session_id, user_id, agent_id, model_alias, message_count, input_tokens, output_tokens, cost_usd, latency_ms, created_at]
      properties:
        id:
          type: integer
        session_id:
          type: string
        user_id:
          type: string
        agent_id:
          type: integer
        model_alias:
          type: string
        message_count:
          type: integer
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost_usd:
          type: number
          format: float
          description: 本次对话成本（美元）
        latency_ms:
          type: integer
        quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 对话质量评分
        user_rating:
          type: integer
          minimum: 1
          maximum: 5
          description: 用户评分（1-5 星）
        created_at:
          type: string
          format: date-time
...